requirements.txt
ipyleaflet
openai
python-dotenv
shiny
shinywidgets